UPSTREAM: soc/intel/elkhartlake: Provide a way to enable real-time tuning
Intel provides a Real-Time Tuning Guide for Elkhart Lake to improve
real-time behaviour of the SoC (see Intel doc #640979). It describes,
amongst knobs for the OS, a couple of firmware settings that need to be
set properly to reduce latencies in all the subsystems. Things like
clock and power gating as well as low power states for peripherals and
buses are disabled in this scenario.

This patch takes the mentioned UEFI parameters from the guide and
translates them to FSP-M and FSP-S parameters. In addition, a chip
config switch guards this tuning which can be selected on mainboard
level if needed.

When this real-time tuning is enabled, the overall system performance
in a real-time environment can be increased by 2-3%.

(cherry picked from commit adbdc5c1bdca62b567c8347e6189427304ee6409)

Original-Change-Id: Ib524ddd675fb3ea270bacf8cd06cb628e895b4b6
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/71233
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Lean Sheng Tan <sheng.tan@9elements.com>
GitOrigin-RevId: adbdc5c1bdca62b567c8347e6189427304ee6409
Change-Id: I6ea1ae0e49de79708b9084fd88aae46e4167da80
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4162188
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Tested-by: Yu-Ping Wu <yupingso@chromium.org>