NOT FOR BRANCHING: exclude SI_DESC area from AP RO verification
This patch removes the SI_DESC area from the range of addresses
covered by AP RO verification to help investigating the reason for
inconsistent ranges' hash value after signing.

It is not clear if the SI_DESC section of the image is modified after
signing, this patch will allow to quickly check this. Any signer
instrumentation modification would take a few days to propagate to
prod signer, whereas this change would result in an updated AP
firmware image produced as soon as the firmware builder/signer runs
after the patch being merged.

Once an image is produced this patch could be reversed.

BRANCH=none
BUG=b:247645824
TEST=none

Change-Id: I3534ffdd81ccad33ed220cf3953d64081534fa8f
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4043446
Reviewed-by: Andrew McRae <amcrae@google.com>