UPSTREAM: vc/amd/fsp/cezanne/FspmUpd: don't use pointers for usb_phy configuration
The size of a pointer changes between a 32 and 64 bit coreboot build. In
order to be able to use a 32 bit FSP in a 64 bit coreboot build, change
the pointer in the UPDs to a uint32_t to always have a 32 bit field in
the UPD for this.

(cherry picked from commit e1f6db512f540363588cb3c3cd77702ad5759821)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I81f3a38344f91cecb4fe5431ed211834e5ed599c
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/69897
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: e1f6db512f540363588cb3c3cd77702ad5759821
Change-Id: Ib4c7e03da3bdcbbcc6d92b7301d11a0f194415cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4092154
Reviewed-by: Reka Norman <rekanorman@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Kangheui Won <khwon@chromium.org>