UPSTREAM: Makefile: Add printall as a NOCOMPILE target
Previously, running "make printall" when there was no .config available,
the system would give an error that printall wasn't a valid target. This
is because it was only in an invalid if clause.  This change adds it to
the other branch of the if clause so it will print out a notice of what
the issue is.

(cherry picked from commit 9fc96407e2279f0178faf27c3b0ae5b3f7341264)

Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Change-Id: I20670ae875be67ac2edf877c53de4702c4fc7c7d
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/70057
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Jakub Czapiga <jacz@semihalf.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
GitOrigin-RevId: 9fc96407e2279f0178faf27c3b0ae5b3f7341264
Change-Id: Ib5673f2e81bef87933fe5df1e6399cca9196f2e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4083220
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Kangheui Won <khwon@chromium.org>
Commit-Queue: Kangheui Won <khwon@chromium.org>