UPSTREAM: vc/amd/fsp/mendocino/FspmUpd: don't use pointers for usb_phy config
The size of a pointer changes between a 32 and 64 bit coreboot build. In
order to be able to use a 32 bit FSP in a 64 bit coreboot build, change
the pointer in the UPDs to a uint32_t to always have a 32 bit field in
the UPD for this.

(cherry picked from commit 7969a5c1b46ba73dce96b3dbd302d59f20331a81)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I419fef73d2881e323487bc7fe641b2ac4041cb17
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/70135
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
GitOrigin-RevId: 7969a5c1b46ba73dce96b3dbd302d59f20331a81
Change-Id: I7613092b27c33609904561ae79b9bb9260a60565
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4092153
Reviewed-by: Kangheui Won <khwon@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Reka Norman <rekanorman@chromium.org>
Tested-by: Reka Norman <rekanorman@chromium.org>