UPSTREAM: vc/amd/fsp/morgana/FspmUpd: don't use pointers for usb_phy config
The size of a pointer changes between a 32 and 64 bit coreboot build. In
order to be able to use a 32 bit FSP in a 64 bit coreboot build, change
the pointer in the UPDs to a uint32_t to always have a 32 bit field in
the UPD for this. Also make sure that the address of the lcl_usb_phy
struct is located below the 4GB boundary, so that the truncation to 32
bits won't result in pointing to a different memory location than
intended. In this error case, which I don't expect to happen, print an
error and write 0 to mcfg->usb_phy_ptr so that the FSP will use its
default values.

(cherry picked from commit bd9ab06808e956135a961566d0644c404b1ab751)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I1394aa6ef5f401e0c7bdd4861f1e28ae46e56e4f
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/70505
Original-Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
Original-Reviewed-by: Matt DeVillier <matt.devillier@amd.corp-partner.google.com>
Original-Reviewed-by: Paul Menzel <paulepanter@mailbox.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: bd9ab06808e956135a961566d0644c404b1ab751
Change-Id: If1531f9fec53a560b9442ab20ea334e78db9ee54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4096697
Reviewed-by: Reka Norman <rekanorman@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Reka Norman <rekanorman@chromium.org>