UPSTREAM: vc/amd/fsp/glinda/FspmUpd: don't use pointers for usb_phy config
The size of a pointer changes between a 32 and 64 bit coreboot build. In
order to be able to use a 32 bit FSP in a 64 bit coreboot build, change
the pointer in the UPDs to a uint32_t to always have a 32 bit field in
the UPD for this.

(cherry picked from commit d3690ee19c2682b4b90e9a2a1ef27609e39d8b24)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I5db2587ff74432a0ce1805d8d7ae76d650693eea
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/70506
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
GitOrigin-RevId: d3690ee19c2682b4b90e9a2a1ef27609e39d8b24
Change-Id: I7e7f959a77d090734825815e3c68d6274d0d22f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4101172
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Reka Norman <rekanorman@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>