UPSTREAM: util/genbuild_h: Update printf %d to %s for sh compatability
When printing a date, genbuild_h is printing it as two digits, using
a leading zero if the value is below 10.

The shells like bash, dash, etc don't fully import the numbers 08 and
09 when using the printf conversion specifier %d.  They apparently
interpret the numbers as octal and only import the leading 0, dropping
the 8 or 9. This isn't an issue for 01 to 07, because those are valid
octal numbers, so %d prints them without an issue.  Because 08 and 09
are not valid octal, various shells return different errors:

Example shell returns for 'printf "%d" 08':
 bash: printf: 08: invalid octal number
 dash: printf: 08: not completely converted
 fish: 008: value not completely converted
 yash: printf: `08' is not a valid integer
 sash: printf: 08: not completely converted

To prevent this, just print all of the values as strings.

zsh just seems to ignore the possibility of the value being octal
and prints the value as a single digit 0-9.

(cherry picked from commit 0110e1abe0ba8e723aa21e8e8fdbc4a27294a6b7)

Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Change-Id: I97b6aa74d74379f6bdc1f0fceecc8002cc36ca09
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/70478
Original-Reviewed-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 0110e1abe0ba8e723aa21e8e8fdbc4a27294a6b7
Change-Id: Id6e75d76cdc60527a731e79f59d4cf343bf777db
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4105198
Reviewed-by: Adam Mills <adamjmills@google.com>
Commit-Queue: Kangheui Won <khwon@chromium.org>
Tested-by: Kangheui Won <khwon@chromium.org>
Reviewed-by: Kangheui Won <khwon@chromium.org>