UPSTREAM: soc/amd/common/block/lpc/espi_util: make eSPI pin setup failure nonfatal
Improve the eSPI pin configuration setup so that a failure in there
won't call die() and stop booting at this point, but return an error to
the caller so that the rest of the eSPI configuration will be skipped.
This will prevent an early boot failure if the EC is missing or the eSPI
interface is in a non-functional state. Also slightly shorten the
function names so that the code still fits into 96 chars.

(cherry picked from commit 84429e092f1c80f6e05e9d15f675923eb043352e)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: Ice2d3a791d6a464eff4fb69d02aeca0bfe580be2
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/70730
Original-Reviewed-by: Martin Roth <martin.roth@amd.corp-partner.google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
Original-Reviewed-by: Nikolai Vyssotski <nikolai.vyssotski@amd.corp-partner.google.com>
GitOrigin-RevId: 84429e092f1c80f6e05e9d15f675923eb043352e
Change-Id: Ia835d525a6796a25f1f03670eae26d814b2a78eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4111185
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Reka Norman <rekanorman@chromium.org>
Reviewed-by: Reka Norman <rekanorman@chromium.org>