UPSTREAM: mb/siemens/mc_apl7: Init I2C controller before PTN3460 is initialized
When PTN3460_EARLY_INIT is selected, the PTN3460 (DP-2-LVDS-bridge) will
be initialized before all devices are initialized. This is necessary to
get a valid EDID data set into the PTN3460 before the graphic controller
is initialized in order to be able to show a splash screen.

For ptn3460_init() to work properly the I2C bus this bridge is connected
to needs to be initialized. As this I2C bus initialization would be done
too late in the normal flow, it needs to be called here explicitly
before ptn3460_init() to initialize the  I2C bus with the needed
conditions. Otherwise the default I2C settings of the controller will be
used which results in a clock rate too high for this mainboard.

Test=Measure I2C bus signals and make sure that the clock is <= 400 kHz.

(cherry picked from commit a1a3be1df84fd76c9195dff182e77fe391101c7f)

Original-Change-Id: I1775fb7c2d29f765224d0e7c7ff9fcd4dbf847c5
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/71226
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: a1a3be1df84fd76c9195dff182e77fe391101c7f
Change-Id: I96cce121376f17af832953456b183b522701932c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4124582
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Yidi Lin <yidilin@chromium.org>
Tested-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Mark Hasemeyer <markhas@google.com>