UPSTREAM: mb/amd/(birman|mayan): Update chromeos.fmd files
Because the EFS is now fixed at 0xff020000, the ChromeOS RO region needs
to be moved to the bottom of the ROM area to cover that space.

The RO Region 6MiB, but you can't actually set 6MiB as RO - it's either
4 or 8MiB, so that's adjusted.  To leave some room for the RW_LEGACY
region, the two RW regions are adjusted to 3MiB each, which should be
plenty.

The GBB region had to be moved from the front of the WP_RO region to the
end to avoid conflicting with the EFS, which needs to be inside the
coreboot cbfs area.

Also get rid of AMD_FWM_POSITION_INDEX.  The FWM position is no longer
needed.

(cherry picked from commit 72f0501881194ed9b222470d82596264bb9c406a)

Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Signed-off-by: Fred Reitberger <reitbergerfred@gmail.com>
Original-Change-Id: I683155ec0f4e6a62d862b9e2fa76af45f4cd5493
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/71770
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
GitOrigin-RevId: 72f0501881194ed9b222470d82596264bb9c406a
Change-Id: Ia1a2b5d96a506d1db235c2fde665f3a959fbeb2a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4186486
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Reka Norman <rekanorman@chromium.org>
Commit-Queue: Reka Norman <rekanorman@chromium.org>