UPSTREAM: mb/starlabs/starbook/adl: Fix the disable wireless CMOS option
The current CMOS option causes Linux to not boot, as the GRUB EFI
loader will report an incorrect parameter.

Update the CMOS option so that the corresponding UPD is changed when
the wireless is set to disable, so that the root port for the wireless
is also disabled.

(cherry picked from commit 6e8c509d7d37d1115fbacccffa3ee7fe2b26e1b6)

Original-Signed-off-by: Sean Rhodes <sean@starlabs.systems>
Original-Change-Id: I607d700319d6a58618ec95b3440e695c82dff196
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/71896
Original-Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 6e8c509d7d37d1115fbacccffa3ee7fe2b26e1b6
Change-Id: I232e276b644133bb347ec90491e802ef16307ddc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4201827
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>