UPSTREAM: vc/siemens/hwilib: Rename 'maxlen' to 'dstsize'
The parameter 'maxlen' can be a bit confusing as it actually is
referring to the size of the destination memory block where the
requested parameter is stored to. Rename it to 'dstsize' and change
the type to size_t to be more clear here.

In addition, add a comment line for this parameter in the description
of the function 'hwilib_get_field()'.

This patch has no impact to the generated binary (checked with timeless
build).

(cherry picked from commit fcff39f0ea47c8c0d82e802058d2ca47c4cc7005)

Original-Change-Id: I572dc0f3ff3d0c177d608332a88991396b82c2fd
Original-Signed-off-by: Werner Zeh <werner.zeh@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/72045
Original-Reviewed-by: Paul Menzel <paulepanter@mailbox.org>
Original-Reviewed-by: Jan Samek <jan.samek@siemens.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Mario Scheithauer <mario.scheithauer@siemens.com>
GitOrigin-RevId: fcff39f0ea47c8c0d82e802058d2ca47c4cc7005
Change-Id: I87c0c84c4e3543fb9875e23ba52f406cb0f0ed50
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4180775
Reviewed-by: Kangheui Won <khwon@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Kangheui Won <khwon@chromium.org>