UPSTREAM: soc/amd/stoneyridge: remove LIDS field from global NVS
Since the LIDS field is only used in the ACPI code and not in the C code
of any mainboard using the Stoneyridge SoC, remove it form the global
NVS and add an ACPI object for this in the DSDT of the mainboards that
use it in their ACPI code. Eventually the LIDS object should probably be
moved to the EC's ACPI code, but that's out of scope for this patch.

TEST=google/liara doesn't show ACPI errors in Linux' dmesg

(cherry picked from commit f56b645f1fe0f70f4b8225ec5f3ee31b31db3d13)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I778c4189607035b4765c6cb8b2e74030dcf9069f
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/72182
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Raul Rangel <rrangel@chromium.org>
Original-Reviewed-by: Matt DeVillier <matt.devillier@gmail.com>
GitOrigin-RevId: f56b645f1fe0f70f4b8225ec5f3ee31b31db3d13
Change-Id: I33264b07c7b876bb720d33ce3fd4b4c88a84c415
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4226806
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>