UPSTREAM: acpi/acpigen: introduce acpigen_write_processor_device
The ACPI PROCESSOR_OP has been deprecated in ACPI 6.0 and dropped in
ACPI 6.4 and is now permanently reserved. As a replacement, DEVICE_OP
with the special HID ACPI0007 should be used instead. This special HID
was introduced in version 3 of the ACPI spec. To have a function to
generate this, acpigen_write_processor_device is introduced. The CPU
index is used as UID which can be assumed to be unique.

(cherry picked from commit 32bba1877ba6d248406a56cd97dce777d499aaa6)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: Ifb0da903a972be134bb3b9071f81b441f60917d1
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/72469
Original-Reviewed-by: Elyes Haouas <ehaouas@noos.fr>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Kyösti Mälkki <kyosti.malkki@gmail.com>
GitOrigin-RevId: 32bba1877ba6d248406a56cd97dce777d499aaa6
Change-Id: Ib9577468ab15ee418cc20f6bffef25e11b1c07b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4212710
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>