UPSTREAM: mb/prodrive/atlas/data.vbt: Fix VBT lane count
Currently there is a problem, where two Displayports are not working. To
be precise: TCP0 and TCP1 (Type-C Port 0/1) are not working.

Setting the lane count of the TCP0 and TCP1 to x1 works fine.
Setting the lane count of the TCP0 and TCP1 to x2 does not work.
Setting the lane count of the TCP0 and TCP1 to x4 does not work.
The reason for that is currently unknown.

This change sets the lane count of the TCP0 and TCP1 Port to x1 length
in the VBT binary.

(cherry picked from commit 6fd406b4a151e5166a273128ec3c7d76b44bdbe0)

Original-Signed-off-by: Maximilian Brune <maximilian.brune@9elements.com>
Original-Change-Id: I182b528275152bf5adcb01a56816afd65674aed3
Original-Signed-off-by: Maximilian Brune <maximilian.brune@9elements.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/72610
Original-Reviewed-by: Lean Sheng Tan <sheng.tan@9elements.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 6fd406b4a151e5166a273128ec3c7d76b44bdbe0
Change-Id: Iecbf66d4830f649ed56281e85e90d8eab064159d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4212719
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>