UPSTREAM: mb/google/brya: Skip locking for GPP_F14 GPIO
This is regarding issues observed on multiple Brya and Nissa
variant such as Skolas and Nivviks. Issue is that once coreboot
sets GPE_EN bit for the GPIO pin and locks it, kernel is not able
to change the control bit. Hence kernel is not able to control the
IRQ on the pin when required.

This issue was root caused to the patch which was setting GPE_EN
bits for the GPIOs before locking.
("intelblocks: Add function to program GPE_EN before GPIO locking")

This patch skips the locking for GPP_F14 to allow kernel to
configure it later during reboot or shutdown as required.

BUG=b:254064671
BRANCH=None
TEST=Shutdown works on Skolas and Brya board with the patch.

(cherry picked from commit f564c8b34abbc7010871cfa426af7ee7261c3fde)

Original-Ref: commit 38b8bf02d820
Original-Signed-off-by: Maulik Vaghela <maulikvaghela@google.com>
Original-Change-Id: I7e4a6ac4668028bcd5fa400b9aa8eccf36a79620
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/72648
Original-Reviewed-by: Caveh Jalali <caveh@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: f564c8b34abbc7010871cfa426af7ee7261c3fde
Change-Id: I98caf4c77924436fcb5a9397b3c08467fbd6f0fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4219522
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Commit-Queue: Paul Fagerburg <pfagerburg@chromium.org>
Reviewed-by: Paul Fagerburg <pfagerburg@chromium.org>