UPSTREAM: soc/amd/phoenix/chipset.cb: rename GPP bridges on device 2
Now that the PCIe ports on device 1 are added, rename the aliases for
the PCIe ports on device 2 to have a common naming scheme. For phoenix
the device alias names are based on the device and function number the
bridge is connected to.

(cherry picked from commit a35b9282cf3bfcc416a17659b846a33ad5458e1c)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I5f5698408019bb9222b599dd78540ca1b187b56d
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/72737
Original-Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: a35b9282cf3bfcc416a17659b846a33ad5458e1c
Change-Id: If20a6149216c0493728f83c895370afd4e5f9dec
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4221885
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>