UPSTREAM: arch/x86/include/cpu: introduce CPUID_FROM_FMS macro
Introduce a macro to get the raw CPUID leaf 1 EAX value from a given set
of CPU family, model and stepping. The processor type in bits 12 and 13
is assumed to be always be zero; at least this is the case for all
CPUIDs that are currently in the coreboot tree. This can be used to
make the device values in the CPU device ID tables easier to read.

(cherry picked from commit 180e414a801935b3557c9a932c87e61acf2c697a)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: Idab77453712b14983b1d02ca365f7924239fc2bf
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/72856
Original-Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
Original-Reviewed-by: Martin Roth <martin.roth@amd.corp-partner.google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 180e414a801935b3557c9a932c87e61acf2c697a
Change-Id: I467690d5d51062888698730a928552e8e7571b0e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4236638
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>