UPSTREAM: mb/amd/birman: Improve missing APCB warning
Move the missing APCB warning to the end of the build and make it stand
out better. Prior to this patch, the warning would appear as one of the
first build messages and easily be missed due to the rest of the build
messages.

TEST=build with and without proper APCBs being found, warning message
appears only when APCB is not found and stands out more

(cherry picked from commit 12931febfdf3f612ba71dce4054817cfbd6caedf)

Original-Signed-off-by: Fred Reitberger <reitbergerfred@gmail.com>
Original-Change-Id: Iabe32636b8e31fe781519533a329a08535bd661a
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/72901
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
GitOrigin-RevId: 12931febfdf3f612ba71dce4054817cfbd6caedf
Change-Id: Iaabe698ba04de7e18213eb17c8b278365113ee32
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4239266
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Nick Vaccaro <nvaccaro@google.com>
Commit-Queue: Nick Vaccaro <nvaccaro@google.com>