UPSTREAM: mb/scaleway/tagada: Drop support
According to the author of the mainboard scaleway/tagada, the mainboard
is not used anymore. Since the mainboard is not publicly available for
purchase and not used anywhere else, the usual deprecation process
of 6 months is not needed.

Thus, to reduce the maintenance overhead for the community, support for
the following components will be removed from the master branch and will
be maintained on the release 4.19 branch. Also, add a note to the 4.20
release notes.

  * Mainboard Scaleway Tagada

(cherry picked from commit c013fa6234f08e22dd3706849936847cf82b5024)

Original-Change-Id: Ifb83b8f2b1dc40cbef657e52c629948dc466ec6e
Original-Signed-off-by: Felix Singer <felixsinger@posteo.net>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/72915
Original-Reviewed-by: Eric Lai <eric_lai@quanta.corp-partner.google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: c013fa6234f08e22dd3706849936847cf82b5024
Change-Id: Ia0ecb31e773ca219bcd2ec2309b8a618e8875975
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4264325
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>
Commit-Queue: Kangheui Won <khwon@chromium.org>
Reviewed-by: Kangheui Won <khwon@chromium.org>