UPSTREAM: Docs/Makefile: change 'which' to 'command -v'
The "which" command is not a posix command. Additionally, if a file
is not found, it outputs "command not found", so when checking to see
if the response is "", that doesn't work.

"command -v" is a posix compliant replacement that doesn't put out any
text if the command is not found, so is more suitable in this case
anyway.

(cherry picked from commit fd8854ec0f51cf1107c8e001fc01da63ceff6110)

Original-Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Change-Id: I22207e818c847d50998f90c9abd55a3cb4bb2ab3
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/72922
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Singer <felixsinger@posteo.net>
GitOrigin-RevId: fd8854ec0f51cf1107c8e001fc01da63ceff6110
Change-Id: I45124ed68c9eaaf172668ed49c3101431f4de9d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4249266
Commit-Queue: Rob Barnes <robbarnes@google.com>
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Rob Barnes <robbarnes@google.com>