UPSTREAM: soc/amd/common/include/psp_efs: rename new PSP directory EFS entry
The EFS entry at offset 0x14 can point to either the first level PSP
directory table or to the PSP combo directory structure that was used
before the introduction of the AMD A/B recovery scheme. This scheme is
not to be confused with the VBOOT scheme. The PSP verstage code checks
if the header this entry points to begins with the PSP_COOKIE, which
indicates the entry is a first level PSP directory table. Due to that,
the EFS entry at offset 0x14 is always expected to point to a PSP
directory table, so rename combo_psp_directory to new_psp_directory to
match the actual usage. This EFS entry that points to the PSP directory
table is called new_psp_directory, since the entry at EFS offset 0x10
was used on some early AMD chips to point to the older PSP directory
table and that one is already called psp_directory. amdfwtool uses the
same naming scheme for those two PSP directory table pointers.

(cherry picked from commit 4bdea41713180865c7a2457a8b20beefead68a59)

Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Change-Id: I10f19ee63f8d422433dba64402d84fd6bb9e0f9e
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/73083
Original-Reviewed-by: Fred Reitberger <reitbergerfred@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
GitOrigin-RevId: 4bdea41713180865c7a2457a8b20beefead68a59
Change-Id: I753326cc30aefdbbd8092686c2ff1a86060df366
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/4274606
Tested-by: CopyBot Service Account <copybot.service@gmail.com>
Reviewed-by: Kangheui Won <khwon@chromium.org>
Commit-Queue: Kangheui Won <khwon@chromium.org>