libpayload/libcbfs: Support file type in unverified_area APIs
These APIs do not currently support obtaining the file type from CBFS.
This is important since it is difficult to run a payload without it.

Expand the APIs to include this.

Note: we might consider reducing the number of combinations, since
there is not a lot of usage of some of these functions.

BUG=b:270062696
BRANCH=none
TEST=build for Brya

Change-Id: I286addab7c435a47e5d22161be84e9fc559de4b9
Signed-off-by: Simon Glass <sjg@chromium.org>