Post-build control of serial
It is annoying to have to create and maintain two completely
different builds of coreboot just to enable or disable the console.
It would be much more convenient to have a 'silent' flag in the
image, which can be updated as needed, without needing to rebuild
coreboot.

Introduce the 'coreboot Control Block' (CCB) which can hold such
settings. It is designed to be available very early in bootblock,
before CBFS is ready. It is able to control the output of the very
first bootblock banner.

The CCB is then placed in cbmem so it is available to other stages.

Provide options in cbfstool to get and set settings in the CCB. This
makes it easy to use this feature.

BUG=none
BRANCH=none
TEST=make (to build coreboot)
$ cbfstool build/coreboot.rom ccb-get -n serial
serial=normal

First try this to see that the bootblock outputs its banner:

$ qemu-system-i386 -bios build/coreboot.rom -nographic |head -5

[NOTE ]  coreboot-4.21 Fri Nov 17 12:09:01 UTC 2023 x86_32
        bootblock starting (log level: 7)...
...

Now set it to silent and try again, to see that the bootblock output is
suppressed, so that the first output shown is the romstage:

$ cbfstool build/coreboot.rom ccb-get -n serial -V silent
serial=silent

$ qemu-system-i386 -bios build/coreboot.rom -nographic
(no output)

Change-Id: I04e946b33035a493e833500351a0483761252613
Signed-off-by: Simon Glass <sjg@chromium.org>