WIP: Proof of concept of post-build control of serial
This shows a possible implementation of a 'silent' console, where the
silence flag can be controlled after the image is built.

This means that we can always build the 'serial' version of firmware,
then create a non-serial version just by changing one byte in the
image, which is much quicker than building everything from scratch.

This is only the start of the full proof-of-concept, so at present it
only affects the coreboot bootblock. We need a Transfer List (TL)
implementation to pass the silence flag through the other stages of
coreboot and into depthcharge.

BUG=b:172341184
BRANCH=none
TEST=On brya
$ ./ccb_serial.sh silent
$ sudo em100 -s -c W25Q256JV -d /path/to/chroot/build/brya/firmware/
	image-felwinter.serial.bin.new -p LOW -r
$ dut-control warm_reset:on sleep:.1 warm_reset:off
See that coreboot produces no output from bootblock (it starts with
ramstage)

Change-Id: I04e946b33035a493e833500351a0483761252613
Signed-off-by: Simon Glass <sjg@chromium.org>