Post-build control of serial
Use the new 'coreboot Control Block' (CCB) to control console output,
including the very first bootblock banner.

Provide a feature in cbfstool to enable and disable the console output.

BUG=none
BRANCH=none
TEST=make (to build coreboot)
$ cbfstool build/coreboot.rom ccb-get -n console
console=loud

First try this to see that the bootblock outputs its banner:

$ qemu-system-i386 -bios build/coreboot.rom -nographic |head -5

[NOTE ]  coreboot-4.21 Fri Nov 17 12:09:01 UTC 2023 x86_32
        bootblock starting (log level: 7)...
...

Now set it to silent and try again, to see that the bootblock output is
suppressed, so that the first output shown is the romstage:

$ cbfstool build/coreboot.rom ccb-get -n console -V silent
console=silent

$ qemu-system-i386 -bios build/coreboot.rom -nographic
(no output)

Change-Id: Ibd867950f117cc6b3dbc582505f3983a0dd714fb
Signed-off-by: Simon Glass <sjg@chromium.org>