Post-build control of serial
It is annoying to have to create and maintain two completely
different builds of coreboot just to enable or disable the console.
It would be much more convenient to have a 'silent' flag in the
image, which can be updated as needed, without needing to rebuild
coreboot.

Introduce the 'Coreboot Control Block' which can hold such settings.
It is designed to be available very early in bootblock, before CBFS
is ready. It is able to control the output of the very first
bootblock banner.

Provide options in cbfstool to get and set settings in the CCB. This
makes it easy to use this feature.

This is only the start of the full proof-of-concept, so at present it
only affects the coreboot bootblock. Future work can implement it for
other stages.

BUG=none
BRANCH=none
TEST=make (to build coreboot)
$ cbfstool build/coreboot.rom ccb-get -n serial
serial=normal

First try this to see that the bootblock outputs its banner:

$ qemu-system-i386 -bios build/coreboot.rom -nographic |head -5

[NOTE ]  coreboot-4.21 Thu Sep 07 02:27:38 UTC 2023 x86_3
        bootblock starting (log level: 7)...
...

Now set it to silent and try again, to see that the bootblock output is
suppressed, so that the first output shown is the romstage:

$ cbfstool build/coreboot.rom ccb-get -n serial -V silent
serial=silent

$ qemu-system-i386 -bios build/coreboot.rom -nographic |head -5

[NOTE ]  coreboot-4.21 Thu Sep 07 02:27:38 UTC 2023 x86_32
        romstage starting (log level: 7)...
...

Change-Id: I04e946b33035a493e833500351a0483761252613
Signed-off-by: Simon Glass <sjg@chromium.org>